Add new models for internal operations
This commit introduces new models in 'internal/models/v1' used for structuring operation data. These models include 'Operation', 'OperationStatus', 'PayloadData', and others. Unit tests have been included in 'operation_test.go' to verify their functionality. This change was required as part of the effort to improve the organization and encapsulation of the data used in internal operations and to provide a clear interface for interaction with this data.