Add in-memory persistence for testing purposes
A method NewBadgerPersistenceWithInMemory was added to the persistence.go file which allows for in-memory database testing. It takes in a context as input and returns a CachePersistence object. If creation fails, an error is returned. It further includes methods to monitor keys and load them. This function builds an in-memory database for use in testing scenarios, since the absence of IO operations speeds up the tests significantly. Changes were reflected in persistence_test.go where now, the in-memory database persistence is created for testing purposes instead of the old file-based persistence. This commit aims to optimise testing performance by avoiding IO operations.