Remove deprecated internal scraper package
The scraper package was removed as it was deprecated and no longer needed in the project. This is part of overall refactoring and clean-up workflows. The go.sum file was also updated to reflect packages that are no longer required. All dependent calls and tests have been updated to match new changes.