Add new persistence tests and a lint workflow
In this commit, the main modifications were the addition of `TestBadgerPersistence_Get` and `TestBadgerPersistence_ListKeys` functions, removal of code redundancy in `persistence_test.go`, and deletion of an unnecessary build directive for Windows. The unused test cases were also removed to aid the maintenance.

A new Github action for linting was also added in `.github/workflows/lint.yaml` which includes actions to setup Go, check out the code, verify dependencies, run checks, check for vulnerabilities and check if the code is properly formatted before initializing tests. This action will ensure that the repository maintains good code quality and standard.

Improvements were necessary for code cleanliness and to ensure uniform code quality standards are maintained in the repository.