Refactor persistence package and update tests
This commit primarily addresses the redundancy and organization of code in the persistence package. Introduced a helper function `badgerInteract` to reduce repeated code blocks and improve ease of reading. The argument 'memory' is added to the `NewBadgerPersistence` function as a flag to change the default database storage mode based on it. Some function names reasoned to be incorrect or misleading were also corrected - for instance, the `DeleteAll` method was renamed to `DropAll` to better represent its functionality.

Corresponding test case in 'persistence_test.go' was updated to accommodate the changes in the source and removed unnecessary ones such as `TestBadgerPersistence_SetStructAsync`.

These updates and modifications overall improve the readability, maintainability and accuracy of the 'persistence' package and its tests.